requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.35.2
torch==2.1.1
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
numpy==1.26.2
Pillow==10.1.0
pydantic==2.5.2
